--- src/main/java/DeadlineException.java
@@ -1,5 +1,0 @@
-class DeadlineException extends Exception {
-    public DeadlineException(String message) {
-        super(message);
-    }
-}
