--- src/main/java/EventException.java
@@ -1,5 +1,0 @@
-class EventException extends Exception {
-    public EventException(String message) {
-        super(message);
-    }
-}
