--- src/main/java/TodoException.java
@@ -1,5 +1,0 @@
-class TodoException extends Exception {
-    public TodoException(String message) {
-        super(message);
-    }
-}
