--- src/main/java/UnknownCommandException.java
@@ -1,5 +1,0 @@
-class UnknownCommandException extends Exception {
-    public UnknownCommandException(String message) {
-        super(message);
-    }
-}
