+++ src/main/java/chatbot/Parser.java
@@ -1,3 +1,9 @@
+package chatbot;
+
+import chatbot.commands.*;
+import chatbot.exceptions.*;
+import chatbot.tasks.*;
+
 public class Parser {
     public static Command parse(String fullCommand) throws UnknownCommandException, TodoException, DeadlineException, EventException {
         String[] inputParts = fullCommand.split(" ", 2);
