+++ src/main/java/chatbot/Ui.java
@@ -1,3 +1,5 @@
+package chatbot;
+
 import java.util.Scanner;
 
 public class Ui {
