+++ src/main/java/chatbot/commands/CommandType.java
@@ -1,3 +1,7 @@
+package chatbot.commands;
+
+import chatbot.tasks.*;
+
 public enum CommandType {
     BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, UNKNOWN;
