+++ src/main/java/chatbot/exceptions/DeadlineException.java
@@ -1,0 +1,7 @@
+package chatbot.exceptions;
+
+public class DeadlineException extends Exception {
+    public DeadlineException(String message) {
+        super(message);
+    }
+}
