+++ src/main/java/chatbot/exceptions/EventException.java
@@ -1,0 +1,7 @@
+package chatbot.exceptions;
+
+public class EventException extends Exception {
+    public EventException(String message) {
+        super(message);
+    }
+}
