+++ src/main/java/chatbot/exceptions/TodoException.java
@@ -1,0 +1,7 @@
+package chatbot.exceptions;
+
+public class TodoException extends Exception {
+    public TodoException(String message) {
+        super(message);
+    }
+}
