+++ src/main/java/chatbot/exceptions/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package chatbot.exceptions;
+
+public class UnknownCommandException extends Exception {
+    public UnknownCommandException(String message) {
+        super(message);
+    }
+}
