+++ src/main/java/chatbot/tasks/Event.java
@@ -1,3 +1,5 @@
+package chatbot.tasks;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
