+++ src/main/java/chatbot/tasks/Task.java
@@ -1,6 +1,8 @@
+package chatbot.tasks;
+
 public class Task {
-    protected String description;
-    protected boolean isDone;
+    public String description;
+    public boolean isDone;
 
     public Task(String description) {
         this.description = description;
