+++ src/main/java/chatbot/tasks/Todo.java
@@ -1,3 +1,5 @@
+package chatbot.tasks;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
@@ -18,5 +20,3 @@ public class Todo extends Task {
         return "[T][" + getStatusIcon() + "] " + description;
     }
 }
-
-
